correct wiki parser
