+++ TreeMap, HashMap/tests/__init__.py
@@ -3,8 +3,9 @@ import unittest
 
 class MyTestCase(unittest.TestCase):
     """"something"""
+
     def test_something(self):
-        """test something"""
+        """tests something"""
         self.assertEqual(True, False)  # add assertion here
